# Add optimization for 64-bit division by constant to mitigate a performance regression on JDK 11 + Graal and GraalVM CE