--- a/jsoniter-scala-core/src/test/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonWriterSpec.scala
+++ b/jsoniter-scala-core/src/test/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonWriterSpec.scala
@@ -535,6 +535,9 @@ class JsonWriterSpec extends WordSpec with Matchers with PropertyChecks {
       check(3.3554432E7f)
       //(1 to Int.MaxValue).par.foreach { n =>
       forAll(minSuccessful(100000)) { (n: Int) =>
+        check(n * Math.pow(10, n & 15).toFloat)
+      }
+      forAll(minSuccessful(100000)) { (n: Int) =>
         val x = java.lang.Float.floatToRawIntBits(n)
         if (java.lang.Float.isFinite(x)) check(x)
       }
@@ -600,6 +603,9 @@ class JsonWriterSpec extends WordSpec with Matchers with PropertyChecks {
       check(1.9430376160308388E16)
       check(-6.9741824662760956E19)
       check(4.3816050601147837E18)
+      forAll(minSuccessful(100000)) { (n: Int) =>
+        check(n * Math.pow(10, n & 31))
+      }
       forAll(minSuccessful(100000)) { (n: Long) =>
         val x = java.lang.Float.floatToRawIntBits(n)
         if (java.lang.Double.isFinite(x)) check(x)